test: pin R's unreadable-tag collection and an arguments-spanning quote
Review follow-ups on the shared fixtures. citation-corpus.json gains a
case quoting across a measure's sources line and arguments block, and
the spec-to-measure builders move into tests/_shared.py and
helper-shared-fixtures.R so both runners share them instead of keeping
an injected-only copy. R gains a per-language test for its side of the
collection divergence the fixture leaves unpinned: R returns an
unrecognized tag value where Python drops it. The shared-fixture
inventory names citation-corpus.json, a Python test stops calling a
tag value an outcome, and escape_attr's docstring states its
double-quoted-attribute-only constraint.
