+++ pkg-r/tests/testthat/fixtures/shared/citation-corpus.json
@@ -1,7 +1,7 @@
 {
   "description": "The trusted text an answer's citations are verified against. The source is tests/shared/citation-corpus.json; the copy under pkg-r/tests/testthat/fixtures/shared/ is generated by scripts/sync-shared.sh. Edit the source and re-run that script.",
   "build_citation_corpus": {
-    "description": "Every case builds a corpus from a context layer's documents, a list of measures, and named data sources whose dictionaries describe a table called `sales`. `expected` names every entry the corpus holds, in order, because matching returns the first entry whose text contains the quote: specific sources are added before the general documentation corpus, so prose that is both a dictionary's and a document's keeps the dictionary's label. `matches` quotes text from the inputs and names the entry that quote must resolve to, or null when nothing may match it. A measure contributes the same block search_pool shows the model, so a quote from that block is citable; measure text itself is pinned by measure-schema.json.",
+    "description": "Every case builds a corpus from a context layer's documents, a list of measures, and named data sources whose dictionaries describe a table called `sales`. `expected` names every entry the corpus holds, in order, because matching returns the first entry whose text contains the quote: specific sources are added before the general documentation corpus, so prose that is both a dictionary's and a document's keeps the dictionary's label. `matches` quotes text from the inputs and names the entry that quote must resolve to, or null when nothing may match it. A measure contributes the same block search_pool shows the model, so a quote from that block is citable; measure text itself is pinned by measure-schema.json, whose spec shape a measure's `arguments` and `injected` fields reuse.",
     "cases": [
       {
         "name": "each kind of trusted text carries its own label and kind",
@@ -146,6 +146,38 @@
           {
             "quote": "Revenue excludes tax collected at checkout.",
             "label": "sales_db dictionary"
+          }
+        ]
+      },
+      {
+        "name": "a quote spans a measure's sources line and arguments block",
+        "docs": [],
+        "measures": [
+          {
+            "name": "region_revenue",
+            "description": "Total revenue for a region.",
+            "arguments": [
+              {
+                "name": "region",
+                "type": "string",
+                "required": true,
+                "description": "The sales region."
+              }
+            ],
+            "injected": ["sales_db"]
+          }
+        ],
+        "sources": [
+          { "name": "sales_db", "dictionary": null },
+          { "name": "returns_db", "dictionary": null }
+        ],
+        "expected": [
+          { "label": "region_revenue definition", "kind": "definition" }
+        ],
+        "matches": [
+          {
+            "quote": "Total revenue for a region.\n\nsources: sales_db\narguments:\n  - region (string, required) The sales region.",
+            "label": "region_revenue definition"
           }
         ]
       }
