+++ tests/shared/README.md
@@ -23,7 +23,7 @@ The Python suite reads this directory directly. The R suite cannot. `testthat` n
 ## What sort of fixtures are here
 
 - **Span names and attributes.** `commons_conversation_turn`, `commons_agent_create`, `commons_data_source_create`, and friends. Also `gen_ai.conversation.id`, `commons.provenance.tag`, and the exact JSON shape of `commons.citation.candidates`. This contract lets the R trajectory reviewer read Python traces. Write it so that it survives the conversation-id ownership moving upstream to shinychat.
-- **The provenance and citation behavior.** The `derive_provenance_tag()` truth table, `collect_appended_tags()` collection from an exchange's turns, `normalize_citation()` input/output pairs, `match_citation()` verdicts including both guards (10-character minimum, only-the-quote-verifies), `parse_commons_citation()` well-formed and malformed bodies. Also the chunk-invariance cases of the streaming scanner. The scanner is a pure chunks-in/string-out function, so it is ideal fixture material.
+- **The provenance and citation behavior.** The `derive_provenance_tag()` truth table, `collect_appended_tags()` collection from an exchange's turns, the `build_citation_corpus()` corpus — which text is citable, under which label, and in which order, since matching returns the first entry holding the quote — `normalize_citation()` input/output pairs, `match_citation()` verdicts including both guards (10-character minimum, only-the-quote-verifies), `parse_commons_citation()` well-formed and malformed bodies. Also the chunk-invariance cases of the streaming scanner. The scanner is a pure chunks-in/string-out function, so it is ideal fixture material.
 - **System prompt rendering.** The template lives in `prompts/`, not here, but what each renderer makes of it does: `prompt-render.json` holds the data a renderer receives and the prompt it must produce. A shared template pins the words, not the rendering.
 - **Where the citation request lands.** `citation-request.json` pins the reminder to the first tool result of a user turn whose output has to be cited: which value shapes it joins and how, and which turns start a new request. The wording lives in `prompts/citation-request.md`, so the fixture uses a stand-in for it.
 - **Handles.** `handles.json` pins the ids a conversation hands out for tool results (`r1`, `r2`, ...), that a value which is not a frame gets one too, the note's opening sentence, and the row cap with the sentence that states it. The opening sentence is pinned as a template with the tool name left as a placeholder, since the tool that reaches a handle is `run_r` in R and `run_python` in Python.
